ControlPlane trait: hoist dual-mode command bodies behind two planes
Phase 3 (final) of the control-plane separation. Adds src/plane.rs — a
#[async_trait(?Send)] ControlPlane trait with 14 verbs derived from the
six dual-mode commands' paired arms — with ServerPlane (wraps client.rs,
connects via require_credentials on the server arm only, with a sync
ServerPlaceholder so pre-login guards keep firing before the login
check) and LocalPlane (owns the Store; ?Send because rusqlite is !Sync —
planes are driven inline under block_on and never spawned).

The six commands collapse to resolve-once-then-verb; their local/server
bodies move verbatim into the plane impls. Printing stays in the impl
where the two planes differ (status blocks, launch recaps) and in the
command where shared (runs table, logs footer).

Domain type plane::Run (+RunLog/RunOrigin) merges the two
run_failure_detail twins — output.rs no longer depends on client::Run.
CLI-facing only: local::model (the up-API wire), the SQLite schema, and
the client.rs DTOs are untouched, verified by byte-identical golden
snapshots of the affected commands on a seeded isolated store.

The plane lives outside src/local/ because local/mod.rs's invariant
(nothing there calls client.rs) would otherwise break. up.rs and
supervise.rs stay off the trait as the deliberate cloud-mirror bridges.
