+++ src/main.rs
@@ -21,6 +21,7 @@ mod jobs;
 #[allow(dead_code)]
 mod local;
 mod output;
+mod plane;
 mod remote;
 mod store;
 mod telemetry;
